+++ plone/protect/views.py
@@ -1,7 +1,8 @@
 from plone.protect.interfaces import IConfirmView
 from Products.Five import BrowserView
-from zope.interface import implements
+from zope.interface import implementer
 
 
+@implementer(IConfirmView)
 class ConfirmView(BrowserView):
-    implements(IConfirmView)
+    pass